% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{durationMapping}
\alias{durationMapping}
\title{Duration Mapping.}
\usage{
durationMapping(x, expMin = 0, expMax = 6, qtrans = NULL, ...)
}
\arguments{
\item{x}{Vector or data frame, series to be mapped}

\item{expMin}{Integer, minimum type is 2^expMin (default: 1 <=> whole note)}

\item{expMax}{Integer, maximum type is 2^expMax (default: 64 <=> 16th note)}

\item{qtrans}{function, quantile transformation to be applied before mapping
For instance data can be "normalized" by using qnorm.}

\item{...}{further arguments to be passed to qtrans.}
}
\value{
A list of duration objects.
}
\description{
Map a series of values into a series of durations
}
\examples{
d <- durationMapping(x=rnorm(100))
}
