% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getNotes}
\alias{getNotes}
\title{Get notes.}
\usage{
getNotes(
  pitches,
  durations = durationMapping(rep(0, length(pitches)), expMin = 4, expMax = 4),
  loudnesses = loudnessMapping(rep(0, length(pitches)), lMin = 89, lMax = 89)
)
}
\arguments{
\item{pitches}{list of pitches (typically created by function pitchMapping)}

\item{durations}{list of durations (typically created by function durationMapping)}

\item{loudnesses}{list of loudnesses (typically created by function loudnessMapping)}
}
\value{
A list of note objects.
}
\description{
Create a series of note objects from lists of pitches / durations / loudnesses
}
\examples{
n <- getNotes(pitches=pitchMapping(x=rnorm(100)))
}
