% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_note.R
\name{note}
\alias{note}
\title{Note constructor.}
\usage{
note(p, d = duration(4), l = 89, tie2next = FALSE, tie2previous = FALSE)
}
\arguments{
\item{p}{Pitch object (step, alter, octave).}

\item{d}{Duration object (type, dot, triplet).}

\item{l}{Numeric (>0), loudness expressed in percentage of a MIDI velocity of 90.
Effective range 0-141 (larger values are clipped).
37: pp, 54: p, 71: mp, 89: mf, 107: f, 124: ff}

\item{tie2next}{Logical, is note tied with next note?.}

\item{tie2previous}{Logical, is note tied with previous note?.}
}
\value{
An object of class 'note'.
}
\description{
Creates a new instance of a 'note' object
}
\examples{
n <- note(p=pitch('Db5'))
}
