% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pitch.R
\name{pitch}
\alias{pitch}
\title{Pitch constructor.}
\usage{
pitch(string)
}
\arguments{
\item{string}{character string comprising:
(i)   one letter in ABCDEFG (step)
(ii)  'b' (flat), '#' (sharp) or '' (no alteration)
(iii) one integer in 0:9 (octave).}
}
\value{
An object of class 'pitch'.
}
\description{
Creates a new instance of a 'pitch' object
}
\examples{
p <- pitch('Db5')
}
