% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pitchMapping}
\alias{pitchMapping}
\title{Pitch Mapping.}
\usage{
pitchMapping(
  x,
  pitches = c("A4", "C5", "D5", "E5", "G5", "A5"),
  qtrans = NULL,
  ...
)
}
\arguments{
\item{x}{Vector or data frame, series to be mapped}

\item{pitches}{Vector of string, pitch scale (default: A minor pentatonic)}

\item{qtrans}{function, quantile transformation to be applied before mapping
For instance data can be "normalized" by using qnorm.}

\item{...}{further arguments to be passed to qtrans.}
}
\value{
A list of pitch objects.
}
\description{
Map a series of values into a series of pitches
}
\examples{
p <- pitchMapping(x=rnorm(100))
}
