% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_score.R
\name{score}
\alias{score}
\title{Score constructor.}
\usage{
score(parts)
}
\arguments{
\item{parts}{List, either a list of measures for a single-part score
or a list of 'parts' (lists of mesures) for a multi-part score}
}
\value{
An object of class 'score'.
}
\description{
Creates a new instance of a 'score' object
}
\examples{
m1 <- measure(number=1,notes=list(note(p=pitch('Db5'),d=duration(2)),
                                  note(p=pitch('B5'),d=duration(2))))
m2 <- measure(number=2,notes=list(note(p=pitch('A5'),d=duration(2)),
                                  note(p=pitch('B5'),d=duration(2))))
s <- score(list(m1,m2))
}
