% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_duration.R
\name{toMXL.duration}
\alias{toMXL.duration}
\title{Duration to MXL (MusicXML)}
\usage{
\method{toMXL}{duration}(x)
}
\arguments{
\item{x}{Duration to be converted.}
}
\value{
A MXL string.
}
\description{
Convert an object of class 'duration' into the corresponding MusicXML chunk
}
\examples{
toMXL(duration(8,dot=TRUE))
}
