% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_measure.R
\name{toMXL.measure}
\alias{toMXL.measure}
\title{Measure to MXL (MusicXML)}
\usage{
\method{toMXL}{measure}(x)
}
\arguments{
\item{x}{measure to be converted.}
}
\value{
A MusicXML string.
}
\description{
Convert an object of class 'measure' into the corresponding MusicXML chunk
}
\examples{
notes=list(note(p=pitch('Db5'),d=duration(2)),note(p=pitch('B5'),d=duration(2)))
m <- measure(number=1,notes=notes)
toMXL.measure(m)
}
