% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_note.R
\name{toMXL.note}
\alias{toMXL.note}
\title{Note to MXL (MusicXML)}
\usage{
\method{toMXL}{note}(x)
}
\arguments{
\item{x}{Note to be converted.}
}
\value{
A MusicXML string.
}
\description{
Convert an object of class 'note' into the corresponding MusicXML chunk
}
\examples{
toMXL.note(note(pitch=pitch('Db5')))
}
