% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pitch.R
\name{toMXL.pitch}
\alias{toMXL.pitch}
\title{Pitch to MXL (MusicXML)}
\usage{
\method{toMXL}{pitch}(x)
}
\arguments{
\item{x}{Pitch to be converted.}
}
\value{
A MusicXML string.
}
\description{
Convert an object of class 'pitch' into the corresponding MusicXML chunk
}
\examples{
toMXL.pitch(pitch('Db5'))
}
