% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toMXL}
\alias{toMXL}
\title{Generic toMXL function}
\usage{
toMXL(o, ...)
}
\arguments{
\item{o}{Object (note, measure or score)}

\item{...}{Optional arguments.}
}
\value{
A MusicXML string.
}
\description{
Generic toMXL function
}
