% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_score.R
\name{toMXL.score}
\alias{toMXL.score}
\title{Score to MXL (MusicXML)}
\usage{
\method{toMXL}{score}(x)
}
\arguments{
\item{x}{score to be converted.}
}
\value{
A MusicXML string.
}
\description{
Convert an object of class 'score' into the corresponding MusicXML chunk
}
\examples{
m1 <- measure(number=1,notes=list(note(p=pitch('Db5'),d=duration(2)),
                                  note(p=pitch('B5'),d=duration(2))))
m2 <- measure(number=2,notes=list(note(p=pitch('A5'),d=duration(2)),
                                  note(p=pitch('B5'),d=duration(2))))
s <- score(list(m1,m2))
toMXL(s)
}
