% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toMXL}
\alias{toMXL}
\title{Generic toMXL function}
\usage{
toMXL(x)
}
\arguments{
\item{x}{Object (note, measure or score)}
}
\value{
A MusicXML string.
}
\description{
Generic toMXL function
}
\examples{
toMXL(note(p=pitch('C5'),d=duration(1),l=107))
}
