% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{writeMXL}
\alias{writeMXL}
\title{writeMXL function}
\usage{
writeMXL(s, file, ...)
}
\arguments{
\item{s}{Score, score object to be written}

\item{file}{Character, destination file}

\item{...}{additional arguments passed to method xml2::write_xml}
}
\description{
Write a score to a musicXML-formatted file
}
